# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "aster-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.19.0"
  },
  "paths": {
...
    "/diagnostics/{session_id}": {
      "get": {
        "tags": [
          "super::routes::status"
        ],
        "operationId": "diagnostics",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Diagnostics zip file",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "500": {
            "description": "Failed to generate diagnostics"
          }
        }
      }
    },
...
  },
...
}